plate-number-service/requirements.txt
fastapi
uvicorn
sqlalchemy
mysql-connector-python
python-dotenv
opencv-python
paddleocr>=2.0.1
paddlepaddle==2.4.2